requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==2.3.2
idna==3.4
multidict==6.0.4
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
typing_extensions==4.8.0
urllib3==1.26.16
yarl==1.9.2
youtube-dl==2021.12.17
